Separate lender vaults to NftLenderVault and TokenLenderVault
