+++ src/components/WalletModal/LenderVaults/index.ts
@@ -1,0 +1,2 @@
+export * from './LenderVaults'
+export * from './helpers'
