+++ src/components/WalletModal/index.ts
@@ -1,4 +1,3 @@
 export * from './WalletModal'
 export * from './hooks'
 export { WalletItem } from './components'
-export * from './helpers'
