+++ src/pages/nftLending/OffersPage/components/OffersTabContent/components/Summary/Summary.tsx
@@ -9,7 +9,7 @@ import { Button } from '@banx/components/Buttons'
 import { EpochProgressBar } from '@banx/components/EpochProgressBar'
 import { StatInfo } from '@banx/components/StatInfo'
 import { DisplayValue } from '@banx/components/TableComponents'
-import { getLenderVaultInfo } from '@banx/components/WalletModal'
+import { getLenderVaultInfo } from '@banx/components/WalletModal/LenderVaults'
 import { BanxSolYieldWarningModal } from '@banx/components/modals'
 
 import { core } from '@banx/api/nft'
