+++ src/pages/tokenLending/OffersTokenPage/components/OfferTokenCard/OfferTokenCard.tsx
@@ -6,6 +6,7 @@ import { Button } from '@banx/components/Buttons'
 import { StatInfo } from '@banx/components/StatInfo'
 import { DisplayValue, createPercentValueJSX } from '@banx/components/TableComponents'
 
+import { convertBondOfferV3ToCore } from '@banx/api/nft'
 import { core } from '@banx/api/tokens'
 import { ChevronDown, Coin, CoinPlus, SOLFilled, USDC, Warning } from '@banx/icons'
 import { useNftTokenType } from '@banx/store/nft'
@@ -27,7 +28,7 @@ const OfferTokenCard: FC<OfferTokenCardProps> = ({ offerPreview, isOpen, onToggl
 
   const onCardClick = () => {
     onToggleCard()
-    setSyntheticOffer(convertToSynthetic(offerPreview.bondOffer, true))
+    setSyntheticOffer(convertToSynthetic(convertBondOfferV3ToCore(offerPreview.bondOffer), true))
   }
 
   return (
