+++ src/pages/tokenLending/OffersTokenPage/components/OffersTokenTabContent/hooks/useTokenOffersPreview.ts
@@ -2,14 +2,18 @@ import { useEffect, useMemo } from 'react'
 
 import { useWallet } from '@solana/wallet-adapter-react'
 import { useQuery } from '@tanstack/react-query'
+import { BondOfferV3 } from 'fbonds-core/lib/fbond-protocol/types'
 import { chain, map, maxBy } from 'lodash'
 
-import { Offer } from '@banx/api/nft'
 import { TokenMarketPreview, core } from '@banx/api/tokens'
 import { useTokenMarketsPreview } from '@banx/pages/tokenLending/LendTokenPage'
 import { useNftTokenType } from '@banx/store/nft'
-import { isOfferNewer, isOptimisticOfferExpired, useTokenOffersOptimistic } from '@banx/store/token'
-import { isOfferClosed } from '@banx/utils'
+import {
+  isOfferNewer,
+  isOptimisticOfferExpired,
+  useTokenOffersOptimisticBN,
+} from '@banx/store/token/useTokenOffersOptimisticBN'
+import { isBondOfferV3Closed } from '@banx/utils/core/tokenOffers'
 
 export const useTokenOffersPreview = () => {
   const { publicKey } = useWallet()
@@ -19,7 +23,7 @@ export const useTokenOffersPreview = () => {
     optimisticOffers,
     remove: removeOffers,
     update: updateOrAddOffer,
-  } = useTokenOffersOptimistic()
+  } = useTokenOffersOptimisticBN()
 
   const { marketsPreview } = useTokenMarketsPreview()
 
@@ -45,9 +49,11 @@ export const useTokenOffersPreview = () => {
     const expiredOffersByTime = optimisticOffers.filter((offer) => isOptimisticOfferExpired(offer))
 
     const optimisticsToRemove = chain(optimisticOffers)
-      .filter(({ offer }) => !isOfferClosed(offer))
+      .filter(({ offer }) => !isBondOfferV3Closed(offer))
       .filter(({ offer }) => {
-        const sameOfferFromBE = userOffers?.find(({ publicKey }) => publicKey === offer.publicKey)
+        const sameOfferFromBE = userOffers?.find(
+          ({ publicKey }) => publicKey?.toBase58() === offer.publicKey.toBase58(),
+        )
         if (!sameOfferFromBE) return false
         const isBEOfferNewer = isOfferNewer(sameOfferFromBE, offer)
         return isBEOfferNewer
@@ -56,7 +62,10 @@ export const useTokenOffersPreview = () => {
 
     if (optimisticsToRemove.length || expiredOffersByTime.length) {
       removeOffers(
-        map([...expiredOffersByTime, ...optimisticsToRemove], ({ offer }) => offer.publicKey),
+        map(
+          [...expiredOffersByTime, ...optimisticsToRemove],
+          ({ offer }) => offer.publicKey?.toBase58(),
+        ),
       )
     }
   }, [data, isFetching, isFetched, optimisticOffers, removeOffers])
@@ -71,12 +80,12 @@ export const useTokenOffersPreview = () => {
     const optimisticUserOffers = optimisticOffers
       .map(({ offer }) => {
         const tokenMarketPreview = marketsPreview.find(
-          ({ marketPubkey }) => marketPubkey === offer.hadoMarket,
+          ({ marketPubkey }) => marketPubkey === offer.hadoMarket?.toBase58(),
         ) as TokenMarketPreview
 
-        return createSynteticOfferPreview(offer, tokenMarketPreview)
+        return createSynteticBondOfferV3Preview(offer, tokenMarketPreview)
       })
-      .filter(({ bondOffer }) => bondOffer.assetReceiver === publicKey?.toBase58())
+      .filter(({ bondOffer }) => bondOffer.assetReceiver.toBase58() === publicKey?.toBase58())
 
     const combinedOffers = [...optimisticUserOffers, ...userOffers]
 
@@ -84,7 +93,7 @@ export const useTokenOffersPreview = () => {
       .groupBy(({ bondOffer }) => bondOffer.publicKey)
       .map((groupedOffers) => maxBy(groupedOffers, ({ bondOffer }) => bondOffer.lastTransactedAt))
       .compact()
-      .filter(({ bondOffer }) => !isOfferClosed(bondOffer))
+      .filter(({ bondOffer }) => !isBondOfferV3Closed(bondOffer))
       .value()
   }, [marketsPreview, data, optimisticOffers, publicKey])
 
@@ -95,15 +104,18 @@ export const useTokenOffersPreview = () => {
   }
 }
 
-const createSynteticOfferPreview = (offer: Offer, tokenMarketPreview: TokenMarketPreview) => {
-  const offerSize = offer.fundsSolOrTokenBalance + offer.bidSettlement
+const createSynteticBondOfferV3Preview = (
+  offer: BondOfferV3,
+  tokenMarketPreview: TokenMarketPreview,
+) => {
+  const offerSize = offer.fundsSolOrTokenBalance.add(offer.bidSettlement).toNumber()
 
   return {
-    publicKey: offer.publicKey,
+    publicKey: offer.publicKey?.toBase58(),
     bondOffer: offer,
     tokenMarketPreview,
     tokenOfferPreview: {
-      publicKey: offer.publicKey,
+      publicKey: offer.publicKey?.toBase58(),
       liquidatedLoansAmount: 0,
       terminatingLoansAmount: 0,
       repaymentCallsAmount: 0,
