+++ src/transactions/nftLending/offers/createClaimLenderVaultTxnData.ts
@@ -5,7 +5,7 @@ import {
   claimPerpetualBondOfferRepayments,
   claimPerpetualBondOfferStakingRewards,
 } from 'fbonds-core/lib/fbond-protocol/functions/perpetual'
-import { LendingTokenType } from 'fbonds-core/lib/fbond-protocol/types'
+import { BondOfferV3, LendingTokenType } from 'fbonds-core/lib/fbond-protocol/types'
 import {
   CreateTxnData,
   SimulatedAccountInfoByPubkey,
@@ -18,7 +18,7 @@ import { BONDS } from '@banx/constants'
 import { banxSol } from '@banx/transactions'
 import { ZERO_BN, isBanxSolTokenType, isOfferStateClosed } from '@banx/utils'
 
-import { parseAccountInfoByPubkey } from '../../functions'
+import { customBNConverter, parseAccountInfoByPubkey } from '../../functions'
 import { sendTxnPlaceHolder } from '../../helpers'
 
 export type CreateClaimLenderVaultTxnDataParams = {
@@ -139,4 +139,12 @@ export const parseClaimLenderVaultSimulatedAccounts = (
   const results = parseAccountInfoByPubkey(accountInfoByPubkey)
 
   return results?.['bondOfferV3']?.[0] as core.Offer
+}
+
+export const parseClaimLenderVaultSimulatedAccountsBN = (
+  accountInfoByPubkey: SimulatedAccountInfoByPubkey,
+) => {
+  const results = parseAccountInfoByPubkey(accountInfoByPubkey, customBNConverter)
+
+  return results?.['bondOfferV3']?.[0] as BondOfferV3
 }
